// generated/src/app/api-models/benefit-accum-weight-detail.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class BenefitAccumWeightDetail {
  accumulatorId: string;
  seqAccumId: number;
  fromValue: string;
  thruValue: string;
  primaryGroup: string;
  secondaryGroup: string;
  weightedAccum: number;
  effectiveDate: Date;
  thruDate: Date;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}